requirements.txt
dataset
diffusers
lightning
matplotlib
numpy
Pillow
scikit-learn
torch
torchvision
transformers